move json response code to own file
